pyproject.toml
[project]
name = "tmpl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adrian Lyjak", email = "adrianlyjak@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "copier>=9.10.2",
    "jinja2>=3.1.6",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "rich>=14.1.0",
]

[project.scripts]
tmpl = "tmpl.cli:cli"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
]
